+++ test/cssCompiler.js
@@ -21,6 +21,8 @@ describe('css preprocessors', () => {
         error: false
       }
     },
+    expressSession: false,
+    secretsDir: 'secrets',
     htmlValidator: {
       enable: false
     },